Now working

+++ stash.sh
@@ -1,45 +1,116 @@
 #!/bin/bash
 
-set -ex
+set -e
 
 OPTIONS="dek:c:D:"
-CIPHER=aes-256-cbc
-MODE=encrypt
-HMAC=sha512
+CIPHER=${STASH_CIPHER:-aes-256-cbc}
+MODE=decrypt
+HMAC=${STASH_HMAC:-sha512}
+KEY=$STASH_KEY
+TMPDIRCLEAN=true
+INFILE=-
+OUTFILE=-
 
-if ( ! getopts $OPTIONS opt); then
+function stash_usage() {
   cat >&2 <<EOF
-Usage:
-  Encrypt: `basename $0` -e [-k key] [-c cipher] [-H hmac] [in] [out]
-  Decrypt: `basename $0` -d -k <key> [in] [out]
-EOF
-  exit $E_OPTERROR;
-fi
+Usage: `basename $0` [OPTION]
+Encrypt or decrypt a file from INFILE (or STDIN) into OUTFILE (or STDOUT).
 
-while getopts $OPTIONS opt; do
-  case $opt in
-    e) MODE=encrypt ;;
-    d) MODE=decrypt ;;
-    c) CIPHER=$OPTARG ;;
-    D) DIGEST=$OPTARG ;;
-    k) KEY=$OPTARG ;;
+  -e, --encrypt
+      Encrypt INFILE and store in OUTFILE, using a provided or
+      generated key.  The key used is written to STDERR.
+  -d, --decrypt
+      Decrypt INFILE (default), using a provided key.
+  -i INFILE, --in INFILE
+      The name of the input file, or - for STDIN (default).
+  -o OUTFILE, --out OUTFILE
+      The name of the output file, or - for STDOUT (default).
+  -k KEY, --key KEY
+      A key of appropriate length in hex format.
+  -c CIPHER, --cipher CIPHER
+      Specify a cipher from one of: aes-256-cbc (default),
+      aes-192-cbc, or aes-128-cbc.
+  -H DIGEST, --hmac DIGEST
+      Specify the digest algorithm to use for the HMAC from one of:
+      sha512 (default), sha384, sha256, sha224, sha
+  -T TMPDIR, --tmpdir TMPDIR
+      Use TMPDIR to (de)construct the stash file.
+
+You may override the defaults using the following variables:
+STASH_KEY, STASH_CIPHER, and STASH_HMAC.
+EOF
+  exit 1
+}
+
+while true ; do
+  case "$1" in
+    -e|--encrypt)
+      MODE=encrypt
+      shift
+      ;;
+
+    -d|--decrypt)
+      MODE=decrypt
+      shift
+      ;;
+
+    -k|--key)
+      KEY=$2
+      shift 2
+      ;;
+
+    -c|--cipher)
+      CIPHER=$2
+      shift 2
+      ;;
+
+    -H|--hmac)
+      HMAC=$2
+      shift 2
+      ;;
+
+    -i|--in)
+      INFILE=$2
+      shift 2
+      ;;
+
+    -o|--out)
+      OUTFILE=$2
+      shift 2
+      ;;
+
+    -T|--tmpdir)
+      TMPDIR=$2
+      TMPDIRCLEAN=false
+      shift 2
+      ;;
+
+    --)
+      shift
+      break
+      ;;
+
+    -*)
+      echo "ERROR: Unknown option: $1" 1>&2
+      stash_usage
+      ;;
+
+    *)
+      break
+      ;;
   esac
 done
 
-shift $((OPTIND-1))
-
-[ "$1" = "--" ] && shift
-
-READ=$1
-WRITE=$2
-
-TMPDIR=$(mktemp -d /tmp/stash-XXXXXXXXXX)
-if [[ -z $READ ]] || [[ $READ = '-' ]]; then
-  READ=/dev/stdin
+if [[ -z $TMPDIR ]]; then
+  TMPDIR=$(mktemp -d /tmp/stash-XXXXXXXXXX)
 fi
 
-if [[ -z $WRITE ]] || [[ $WRITE = '-' ]]; then
-  WRITE=/dev/stdout
+if [[ -z $INFILE ]] || [[ $INFILE = '-' ]]; then
+  INFILE=/dev/stdin
+fi
+
+if [[ -z $OUTFILE ]] || [[ $OUTFILE = '-' ]]; then
+  OUTFILE=/dev/stdout
 fi
 
 function error() {
@@ -48,153 +119,194 @@ function error() {
   exit 1
 }
 
-function stash_make_nonce() {
-  SIZE=${1:-16}
-  RANDOM=$(( $SIZE - 16 ))
-  if [[ $RANDOM -gt 0 ]]; then
-    #RANDOM=$(openssl rand -hex $RANDOM)
-    true
-  else
-    RANDOM=""
+function stash_nonce() {
+  SIZE=$(( $1 + 0 ))
+  LENGTH=$(( $1 * 2 ))
+
+  if [[ $SIZE -lt 16 ]]; then
+    error "Unable to create nonce shorter than 16 bytes"
   fi
-  RET=$(date +"%s %N $RANDOM" | awk '{printf "%08x%08x%s\n", $1, $2, $3}' | cut -c 1-$SIZE)
+
+  RET=$(date +"%s %N" | awk '{printf "%08x%08x\n", $1, $2}')
+
+  EXTRA=$(( $LENGTH - ${#RET} ))
+  if [[ $EXTRA -gt 0 ]]; then
+    RET="$RET$(openssl rand -hex $(( $EXTRA / 2 )))"
+  fi
 }
 
-function stash_hmac() {
+function openssl_hmac() {
   local DIGEST=$1
   local KEY=$2
-  local FILE=$3
-  local CMD
 
   case $DIGEST in
     sha|sha1|sha-1|SHA|SHA1|SHA-1)
-      CMD="-sha1"
       DIGEST=sha1
       ;;
 
     sha224|sha-224|SHA224|SHA-224)
-      CMD="-sha224"
       DIGEST=sha224
       ;;
 
     sha256|sha-256|SHA256|SHA-256)
-      CMD="-sha256"
       DIGEST=sha256
       ;;
 
     sha384|sha-384|SHA384|SHA-384)
-      CMD="-sha384"
       DIGEST=sha384
       ;;
 
     sha512|sha-512|SHA512|SHA-512)
-      CMD="-sha512"
       DIGEST=sha512
       ;;
     *)
       error "Unknown digest: $DIGEST" ;;
   esac
 
-  set -- $(openssl dgst -r -hmac $KEY $CMD < $FILE)
-  RET="$DIGEST $1"
+  set -- $(openssl dgst -r -hmac $KEY -$DIGEST)
+  echo "$DIGEST $1"
 }
 
 function stash_cipher_params() {
+  local CIPHER
+  local KEYSIZE
+  local IVSIZE
+
   case $1 in
     aes128|aes-128-cbc|AES-128-CBC)
-      RET="aes-128-cbc 8"
+      CIPHER=aes-128-cbc
+      KEYSIZE=16
+      IVSIZE=16
       ;;
 
     aes192|aes-192-cbc|AES-192-CBC)
-      RET="aes-192-cbc 24"
+      CIPHER=aes-192-cbc
+      KEYSIZE=24
+      IVSIZE=16
       ;;
 
     aes256|aes-256-cbc|AES-256-CBC)
-      RET="aes-256-cbc 16"
+      CIPHER=aes-256-cbc
+      KEYSIZE=32
+      IVSIZE=16
       ;;
 
     *)
       error "Unknown cipher: $CIPHER"
   esac
+
+  if [[ -z $IVSIZE ]]; then
+    IVSIZE=$KEYSIZE
+  fi
+
+  echo "$CIPHER $KEYSIZE $IVSIZE"
+}
+
+function check_key_iv() {
+  local CIPHER=$1
+  local KEY=$2
+  local IV=$3
+
+  set -- $(stash_cipher_params $CIPHER)
+  local KEYSIZE=$2
+  local IVSIZE=$3
+
+  if [[ $(( ${#KEY} / 2 )) != $KEYSIZE ]]; then
+    error "The key is the wrong size: $(( ${#KEY} / 2) )) != $KEYSIZE"
+  fi
+
+  if [[ $(( ${#IV} / 2 )) != $IVSIZE ]]; then
+    error "The iv is the wrong size: $(( ${#IV} / 2 )) != $IVSIZE"
+  fi
+}
+
+function openssl_encrypt() {
+  local CIPHER=$1
+  local KEY=$2
+  local IV=$3
+
+  openssl enc -e -$CIPHER -K $KEY -iv $IV
+}
+
+function openssl_decrypt() {
+  local CIPHER=$1
+  local KEY=$2
+  local IV=$3
+
+  openssl enc -d -$CIPHER -K $KEY -iv $IV
 }
 
 function stash_encrypt() {
-  stash_cipher_params $CIPHER
-  set -- $RET
+  set -- $(stash_cipher_params $CIPHER)
   CIPHER=$1
-  BSIZE=$2
+  KEYSIZE=$2
+  IVSIZE=$3
 
   if [[ -z $KEY ]]; then
     if [[ $STASH_KEY ]]; then
       KEY=$STASH_KEY
     else
-      KEY=$(openssl rand -hex $BSIZE)
+      KEY=$(openssl rand -hex $KEYSIZE)
     fi
   fi
 
-  if [[ ${#KEY} -ne $(( $BSIZE * 2 )) ]]; then
-    error "Supplied key is not the correct length: ${#KEY} != $BSIZE"
-  fi
-
-  stash_make_nonce
+  stash_nonce $IVSIZE
   IV=$RET
 
-  cat $READ > $TMPDIR/unencrypted
+  check_key_iv $CIPHER $KEY $IV
 
-  stash_hmac $HMAC $KEY $TMPDIR/unencrypted
-  HMAC_DATA=$RET
-  
-  openssl enc -e -$CIPHER -K $KEY -iv $IV -in $TMPDIR/unencrypted -out $TMPDIR/encrypted || error "Unable to encrypt data"
-  rm -f $TMPDIR/unencrypted
+  HMAC_DATA=$(tee < $INFILE >(openssl_encrypt $CIPHER $KEY $IV > $TMPDIR/encrypted) | openssl_hmac $HMAC $KEY)
 
-  cat > $TMPDIR/control <<EOF
-Cipher: $CIPHER
-IV: $IV
-HMAC: $HMAC_DATA
-EOF
-  
-  echo "Key: $KEY" 1>&2
+  echo "Timestamp: $(date -u --iso-8601=seconds)" > $TMPDIR/control
+  echo "Cipher: $CIPHER" >> $TMPDIR/control
+  echo "IV: $IV" >> $TMPDIR/control
+  echo "HMAC: $HMAC_DATA" >> $TMPDIR/control
+
   cat $TMPDIR/control 1>&2
+  echo "Key: $KEY" 1>&2
 
-  cat > $TMPDIR/manifest <<EOF
-control
-encrypted
-EOF
-  
-  #tar -C $TMPDIR -c control encrypted | base64
-  (cd $TMPDIR && cat manifest | cpio -o > $WRITE)
+  echo "control" > $TMPDIR/manifest
+  echo "encrypted" >> $TMPDIR/manifest
+
+  (cd $TMPDIR && cat manifest | cpio --quiet -o) > $OUTFILE
 }
 
 function stash_decrypt() {
   if [[ -z $KEY ]]; then
-    error "Unable to decrypt without supplying a key"
+    error "Unable to decrypt without providing a key"
   fi
-  
-  (cd $TMPDIR && cpio -i) < $READ
-  set -- $(grep "^Cipher:" < $TMPDIR/control)
-  CIPHER=$2
 
-  set -- $(grep "^HMAC:" < $TMPDIR/control)
-  HMAC=$2
+  (cd $TMPDIR && cpio --quiet -i) < $INFILE
 
-  set -- $(grep "^IV:" < $TMPDIR/control)
-  IV=$2
+  while read line; do
+    set -- $line
+    case $1 in
+      Cipher:)
+        CIPHER=$2
+        ;;
 
-  stash_cipher_params $CIPHER
-  set -- $RET
-  CIPHER=$1
-  BSIZE=$2
+      HMAC:)
+        HMAC=$2
+        HMAC_OK=$3
+        ;;
 
-  openssl enc -d -$CIPHER -K $KEY -iv $IV -in $TMPDIR/encrypted -out $TMPDIR/unencrypted || error "Unable to encrypt data"
-  rm -rf $TMPDIR/encrypted
+      IV:)
+        IV=$2
+        ;;
+    esac
+  done < $TMPDIR/control
 
-  stash_hmac $HMAC $KEY $TMPDIR/unencrypted
-  set -- $RET
-  ACTUAL_HMAC=$2
+  check_key_iv $CIPHER $KEY $IV
 
-  cat $TMPDIR/unencrypted > $WRITE
-  
-  ls -l $TMPDIR
+  HMAC_DATA=$(openssl_decrypt $CIPHER $KEY $IV < $TMPDIR/encrypted | openssl_hmac $HMAC $KEY)
+  set -- $HMAC_DATA
+  HMAC_OUT=$2
+
+  if [[ $HMAC_OUT != $HMAC_OK ]]; then
+    error "HMAC verification failed!  Are you using the right key?"
+  fi
+
+  openssl_decrypt $CIPHER $KEY $IV < $TMPDIR/encrypted > $OUTFILE
 }
 
 case $MODE in
@@ -202,7 +314,8 @@ case $MODE in
   decrypt) stash_decrypt ;;
 esac
 
-rm -rf $TMPDIR
+if [[ $TMPDIRCLEAN = true ]]; then
+  rm -rf $TMPDIR
+fi
 
 exit 0
-
